♻️refactor: DTO 패키지 구조 변경으로 인한 자동 수정
