+++ src/main/java/checkmo/domain/book/facade/BookQueryFacade.java
@@ -1,6 +1,6 @@
 package checkmo.domain.book.facade;
 
-import checkmo.domain.book.dto.BookResponseDTO;
+import checkmo.domain.book.web.dto.BookResponseDTO;
 import checkmo.global.dto.BookSharedDTO;
 
 /**
